requirements.txt
paramiko
python-dotenv 